Remove useless window.setTimeout call
Well, the code has been passing the *value* of
`sendMetaInfoToExtension()` to `window.setTimeout` for the past 5
years. That value is `undefined`, which evidently gets subjected to an
[approximation of `eval`][1]. There don't seem to be any observable
effects of this other than triggering a CSP violation when the page's
`script-src` policy doesn't include `unsafe-eval`, so just call
`sendMetaInfoToExtension()` and don't set the timeout.

[1]: https://html.spec.whatwg.org/multipage/timers-and-user-prompts.html#timers:report-the-exception